added up to 1000 batch limit for db_migration api

+++ lambda/db_migration/db_migration.py
@@ -38,10 +38,10 @@ def _read_csv(payload_body: str):
     return df
 
 
-def _send_data_to_mysql(data_df: pd.DataFrame, table: str, schema: str):
+def _send_data_to_mysql(data_df: pd.DataFrame, table: str, schema: str, append:bool=False):
     con = wr.mysql.connect(secret_id=secret_name)
     tables = wr.mysql.read_sql_query(sql="show tables;", con=con)
-    if table in tables.iloc[:, 0].to_list():
+    if table in tables.iloc[:, 0].to_list() and not append:
         with con.cursor() as cursor:
             cursor.execute(f"DROP TABLE {table}")
     wr.mysql.to_sql(df=data_df, table=table, schema=schema, con=con)
@@ -52,19 +52,30 @@ def lambda_handler(event, context):
     print("Received event:", json.dumps(event))
 
     res = {"statusCode": 200, "headers": {"Content-Type": "*/*"}}
-    res["body"] = f"Data loaded succesfully!"
+    res["body"] = f"Default behavior, nothing was done."
 
     chk_params = _check_input_parameters(event.get("queryStringParameters"))
 
     if event["headers"].get("Content-Type") == "text/csv" and chk_params == None:
         df = _read_csv(event.get("body"))
-        table = event.get("queryStringParameters").get("table")
-        schema = event.get("queryStringParameters").get("schema")
 
-        print("Table: ", table)
-        print("Schema: ", schema)
+        if df.shape[0] > 1000:
+            res = {
+                "statusCode": 400,
+                "body": f"Batch insert available for files from 1 up to 1000 rows. The file provided has {df.shape[0]} rows.",
+                "headers": {"Content-Type": "*/*"},
+            }
+        else:
+            table = event.get("queryStringParameters").get("table")
+            schema = event.get("queryStringParameters").get("schema")
+            append = event.get("queryStringParameters").get("append",False)
 
-        _send_data_to_mysql(data_df=df, table=table, schema=schema)
+            print("Table: ", table)
+            print("Schema: ", schema)
+
+            _send_data_to_mysql(data_df=df, table=table, schema=schema, append=append)
+            res = {"statusCode": 200, "headers": {"Content-Type": "*/*"}}
+            res["body"] = f"Data loaded succesfully!"
 
     elif chk_params != None:
         res = chk_params
